+++ src/ska_pss_protest/product_tests/test_rfim_sps.py
@@ -12,7 +12,13 @@ from xml.etree import ElementTree as et
 import pytest
 from pytest_bdd import given, parsers, scenarios, then, when
 
-from ska_pss_protest import Cheetah, SpCcl, VectorPull, VHeader
+from ska_pss_protest import (
+    Cheetah,
+    DedispersionPlanSelect,
+    SpCcl,
+    VectorPull,
+    VHeader,
+)
 
 # pylint: disable=W0621,W0212,C0116,C0103,C0301
 
@@ -159,8 +165,8 @@ def set_rfim_iqrm_zdot(config, threshold, radius):
     config("rfim/rfim_zdot/active", "true")
 
 
-@when("An SPS pipeline runs")
-def run_cheetah(context, config, pytestconfig):
+@when(parsers.parse("An SPS pipeline runs using {dedispersion_plan}"))
+def run_cheetah(context, config, pytestconfig, dedispersion_plan):
     """
     Add SpCcl output directory to config and
     run cheetah
@@ -172,8 +178,18 @@ def run_cheetah(context, config, pytestconfig):
     # Set number of samples in dedispersion buffer
     context["dd_samples"] = 131072
     root_tree = config("ddtr/dedispersion_samples", str(context["dd_samples"]))
-    # read the root tree for trial widths
 
+    dd_plan = DedispersionPlanSelect()
+
+    for segment in dd_plan.select(dedispersion_plan):
+        dedispersion = et.Element("dedispersion")
+
+        for key, value in segment.items():
+            child = et.SubElement(dedispersion, key)
+            child.text = str(value)
+        root_tree.find("ddtr").append(dedispersion)
+
+    # read the root tree for trial widths
     widths_element = root_tree.find("sps/klotski/widths")
     if widths_element is not None and widths_element.text:
         trial_widths = [int(r) for r in widths_element.text.strip().split(",")]